requirements.txt
# Dependencias principales para OCR matemático
PyQt5==5.15.10
easyocr==1.7.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# Matemáticas y demostraciones
matplotlib==3.7.2
sympy==1.12
torch==2.0.1
torchvision==0.15.2
z3-solver==4.12.2
networkx==3.1

# Procesamiento de texto
nltk==3.8.1

# Nota: Tesseract eliminado - solo EasyOCR para manuscritos matemáticos
